PlayerInteractables print their ID every frame
This was done for testing purposes to figure out why their IDs weren't being set correctly when loading. Going off of the print statements, it seems that a PlayerInteractable's ID is reset after the player fully loads into the level, but that's just a theory. John will hopefully look into this more tomorrow