rope reset  within range
another reason ropes sometimes werent pulling was the end point was too far when the rope was initially reset on pickup. so now im reseting the rigid rope again once the slot is within the range to make it pull properly.  minor tweaks to far ropes start positions and tween time.